test: split response event terminal coverage
